+++ enterprise/frontend/src/embedding-sdk/lib/theme/default-component-theme.ts
@@ -2,6 +2,7 @@ import { merge } from "icepick";
 
 import type { MetabaseComponentTheme } from "embedding-sdk";
 import { EMBEDDING_SDK_ROOT_ELEMENT_ID } from "embedding-sdk/config";
+import type { DeepPartial } from "embedding-sdk/types/utils";
 import type { MantineThemeOverride } from "metabase/ui";
 
 export const DEFAULT_SDK_FONT_SIZE = 14;
@@ -28,6 +29,16 @@ export const FONT_SIZES = {
  * such as charts, data tables and popovers.
  */
 export const DEFAULT_METABASE_COMPONENT_THEME: MetabaseComponentTheme = {
+  collectionBrowser: {
+    breadcrumbs: {
+      expandButton: {
+        textColor: "var(--mb-color-text-medium)",
+        backgroundColor: "var(--mb-color-bg-light)",
+        hoverTextColor: "var(--mb-color-text-white)",
+        hoverBackgroundColor: "var(--mb-color-brand)",
+      },
+    },
+  },
   dashboard: {
     backgroundColor: "var(--mb-color-bg-white)",
     card: {
@@ -64,23 +75,32 @@ export const DEFAULT_METABASE_COMPONENT_THEME: MetabaseComponentTheme = {
  * Default theme options, with overrides specific to the
  * Embedding SDK environment to provide nicer defaults.
  */
-export const DEFAULT_EMBEDDED_COMPONENT_THEME: MetabaseComponentTheme = merge(
-  DEFAULT_METABASE_COMPONENT_THEME,
-  {
-    table: {
-      cell: {
-        fontSize: FONT_SIZES.tableCell.em,
-        backgroundColor: "bg-white",
-      },
+export const DEFAULT_EMBEDDED_COMPONENT_THEME: MetabaseComponentTheme = merge<
+  MetabaseComponentTheme,
+  DeepPartial<MetabaseComponentTheme>
+>(DEFAULT_METABASE_COMPONENT_THEME, {
+  table: {
+    cell: {
+      fontSize: FONT_SIZES.tableCell.em,
+      backgroundColor: "bg-white",
     },
-    cartesian: {
-      label: { fontSize: FONT_SIZES.label.em },
-      goalLine: {
-        label: { fontSize: FONT_SIZES.goalLabel.em },
+  },
+  cartesian: {
+    label: { fontSize: FONT_SIZES.label.em },
+    goalLine: {
+      label: { fontSize: FONT_SIZES.goalLabel.em },
+    },
+  },
+  collectionBrowser: {
+    breadcrumbs: {
+      expandButton: {
+        backgroundColor: "transparent",
+        hoverTextColor: "var(--mb-color-text-white)",
+        hoverBackgroundColor: "var(--mb-color-brand)",
       },
     },
   },
-);
+});
 
 export const EMBEDDING_SDK_COMPONENTS_OVERRIDES: MantineThemeOverride["components"] =
   {