+++ frontend/src/metabase/admin/performance/utils.tsx
@@ -1,4 +1,4 @@
-import _, { memoize } from "underscore";
+import { memoize } from "underscore";
 import type { SchemaObjectDescription } from "yup/lib/schema";
 
 import {
@@ -9,14 +9,14 @@ import {
 import { isNullOrUndefined } from "metabase/lib/types";
 import { PLUGIN_CACHING } from "metabase/plugins";
 import type {
+  CacheConfig,
+  CacheStrategy,
+  CacheStrategyType,
   CacheableModel,
-  Config,
   ScheduleDayType,
   ScheduleFrameType,
   ScheduleSettings,
   ScheduleType,
-  Strategy,
-  StrategyType,
 } from "metabase-types/api";
 
 import { defaultMinDurationMs, rootId } from "./constants/simple";
@@ -188,7 +188,7 @@ export const getFrequencyFromCron = (cron: string) => {
 
 export const isValidStrategyName = (
   strategy: string,
-): strategy is StrategyType => {
+): strategy is CacheStrategyType => {
   const { strategies } = PLUGIN_CACHING;
   const validStrategyNames = new Set(Object.keys(strategies));
   return validStrategyNames.has(strategy);
@@ -198,7 +198,7 @@ export const getLabelString = (label: StrategyLabel, model?: CacheableModel) =>
   typeof label === "string" ? label : label(model);
 
 export const getShortStrategyLabel = (
-  strategy?: Strategy,
+  strategy?: CacheStrategy,
   model?: CacheableModel,
 ) => {
   const { strategies } = PLUGIN_CACHING;
@@ -215,7 +215,7 @@ export const getShortStrategyLabel = (
   }
 };
 
-export const getFieldsForStrategyType = (strategyType: StrategyType) => {
+export const getFieldsForStrategyType = (strategyType: CacheStrategyType) => {
   const { strategies } = PLUGIN_CACHING;
   const strategy = strategies[strategyType];
   const validationSchemaDescription =
@@ -226,10 +226,10 @@ export const getFieldsForStrategyType = (strategyType: StrategyType) => {
 };
 
 export const translateConfig = (
-  config: Config,
+  config: CacheConfig,
   direction: "fromAPI" | "toAPI",
-): Config => {
-  const translated: Config = { ...config };
+): CacheConfig => {
+  const translated: CacheConfig = { ...config };
 
   // If strategy type is unsupported, use a fallback
   if (!isValidStrategyName(translated.strategy.type)) {
@@ -253,7 +253,7 @@ export const translateConfig = (
   return translated;
 };
 
-export const translateConfigFromAPI = (config: Config): Config =>
+export const translateConfigFromAPI = (config: CacheConfig): CacheConfig =>
   translateConfig(config, "fromAPI");
-export const translateConfigToAPI = (config: Config): Config =>
+export const translateConfigToAPI = (config: CacheConfig): CacheConfig =>
   translateConfig(config, "toAPI");